+++ CHANGELOG.rst
@@ -32,6 +32,7 @@ srt-py Change Log
 * Logging messages to beam switch
 * Warning message when angle out of bounds during n-point scan and beam switch
 * Log message at the end of n-point scan and beam switch
+* Beam switch graph
 
 **Changed:**